+++ R/lingmatch.R
@@ -30,10 +30,9 @@
 #'   \item If a \strong{character vector}, this will be processed in the same way as \code{input}.
 #'   \item If a \strong{vector}, either (a) logical or factor-like (having n levels < length) and of the same length as
 #'     \code{nrow(input)}, or (b) numeric or logical of length less than \code{nrow(input)}, this will be used to
-#'     select a subset of \code{input} (e.g., \code{comp = 1:10} would treat the first 10 rows of \code{input} as the
-#'     comparison; \code{comp = type == 'prompt'} would make a logical vector identifying prompts, assuming "type" was
-#'     the name of a column in \code{data}, or a variable in the global environment, and the value "prompt" marked the
-#'     prompts).
+#'     select a subset of \code{input} (e.g., \code{1:10} would treat the first 10 rows of \code{input} as the
+#'     comparison; \code{lingmatch(text, type == 'prompt', data)} would use the texts in the \code{text} column
+#'     identified by the \code{type} column as the comparison).
 #'   \item If a \strong{matrix-like object} (having multiple rows and columns), or a named vector, this will
 #'     be treated as a sort of dtm, assuming there are common (column) names between \code{input} and
 #'     \code{comp} (e.g., if you had prompt and response texts that were already processed separately).
@@ -109,7 +108,7 @@
 #'
 #' Niederhoffer, K. G., & Pennebaker, J. W. (2002). Linguistic style matching in social interaction.
 #'   \emph{Journal of Language and Social Psychology, 21}, 337-360.
-#' @seealso For a general text processing function, see \code{\link{lma_process}}.
+#' @seealso For a general text processing function, see \code{\link{lma_process}()}.
 #' @return A list with processed components of the input, information about the comparison, and results of
 #' the comparison:
 #' \itemize{
@@ -265,7 +264,6 @@ lingmatch <- function(input = NULL, comp = mean, data = NULL, group = NULL, ...,
   input <- if (is.character(input) && all(input %in% colnames(data))) data[, input] else gd(opt$input, data)
   if (!missing(group) && is.data.frame(input)) input <- as.matrix(input[, vapply(input, is.numeric, TRUE)])
   rx <- NROW(input)
-  cx <- NCOL(input)
   # comp
   if (!missing(comp)) {
     comp <- gd(opt$comp, if (missing(comp.data)) if (is.call(opt$comp)) NULL else data else comp.data)
@@ -304,7 +302,6 @@ lingmatch <- function(input = NULL, comp = mean, data = NULL, group = NULL, ...,
     if (sum(ck) / length(ck) > .75) {
       inp$dict <- NULL
       if (any(!ck)) dn <- dn[ck]
-      cx <- length(dn)
       input <- input[, dn]
       do.wmc <- FALSE
       if (!missing(comp) && any(class(comp) %in% c("matrix", "data.frame")) && all(dn %in% colnames(comp))) {