+++ R/report_term_matches.R
@@ -21,9 +21,17 @@
 #' @param bysentence Logical; if \code{TRUE}, will split \code{text} into sentences, and only
 #' consider unique sentences.
 #' @param as_string Logical; if \code{FALSE}, returns matches as tables rather than a string.
+#' @param term_map_freq Proportion of terms to include when using the term map as a source
+#' of terms. Applies when \code{text} is not specified.
+#' @param term_map_spaces Number of spaces in which a term has to appear to be included.
+#' Applies when \code{text} is not specified.
 #' @param outFile File path to write results to, always ending in \code{.csv}.
 #' @param space_dir Directory from which \code{space} should be loaded.
 #' @param verbose Logical; if \code{FALSE}, will not display status messages.
+#' @family Dictionary functions
+#' @seealso For a more complete assessment of dictionaries, see \code{\link{dictionary_meta}()}.
+#'
+#' Similar information is provided in the \href{https://miserman.github.io/dictionary_builder/}{dictionary builder} web tool.
 #' @note
 #' Matches are extracted for each term independently, so they may not align with some implementations
 #' of dictionaries. For instance, by default \code{\link{lma_patcat}} matches destructively, and sorts
@@ -65,14 +73,29 @@
 
 report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
                                 parse_phrases = TRUE, tolower = TRUE, punct = TRUE, special = TRUE,
-                                as_terms = FALSE, bysentence = FALSE, as_string = TRUE, outFile = NULL,
+                                as_terms = FALSE, bysentence = FALSE, as_string = TRUE,
+                                term_map_freq = 1, term_map_spaces = 1, outFile = NULL,
                                 space_dir = getOption("lingmatch.lspace.dir"), verbose = TRUE) {
-  if (missing(dict)) stop("dict must be specified", .call = FALSE)
+  if (missing(dict)) stop("dict must be specified", call. = FALSE)
+  collapsed_terms <- FALSE
   if (is.null(text)) {
-    text <- rownames(select.lspace(dir = space_dir, get.map = TRUE)$term_map)
-    as_terms <- TRUE
+    term_map <- select.lspace(dir = space_dir, get.map = TRUE)$term_map
+    if (is.null(term_map)) {
+      stop(
+        "term map not found; specify `space_dir` or provide text",
+        call. = FALSE
+      )
+    }
+    if (term_map_freq > 0 && term_map_freq < 1) {
+      term_map <- term_map[seq(1, ceiling(nrow(term_map) * term_map_freq)), ]
+    }
+    if (term_map_spaces > 0 && term_map_spaces < ncol(term_map)) {
+      term_map <- term_map[rowSums(term_map != 0) >= term_map_spaces, ]
+    }
+    collapsed_terms <- as_terms <- TRUE
+    text <- paste(rownames(term_map), collapse = "  ")
   }
-  if (is.null(text) && is.null(space)) stop("either text or space must be specified", .call = FALSE)
+  if (is.null(text) && is.null(space)) stop("either text or space must be specified", call. = FALSE)
   st <- proc.time()[[3]]
   if (!is.null(text) && !as_terms) {
     if (verbose) cat("preparing text (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
@@ -92,16 +115,7 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
       if (glob) glob <- !any(grepl("(?:\\\\\\w|[].})])\\*$", terms$term))
     }
   }
-  space_name <- NULL
-  if (is.null(text)) {
-    if (is.character(space)) {
-      space_name <- space
-      if (verbose) cat("preloading space (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
-      space <- lma_lspace(space, dir = space_dir)
-    }
-    text <- rownames(space)
-  }
-  rawtext <- is.null(space_name) && !as_terms
+  rawtext <- !as_terms || collapsed_terms
   terms$regex <- to_regex(list(terms$term), TRUE, glob)[[1]]
   terms <- terms[!is.na(terms$regex) & terms$regex != "", ]
   terms$regex <- if (rawtext) paste0("\\b", terms$regex, "\\b") else paste0("^", terms$regex, "$")
@@ -121,18 +135,17 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
   if (verbose) cat("extracting matches (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
   matches <- extract_matches(terms$regex, text, rawtext)
   has_space <- FALSE
+  space_name <- NULL
   if (!is.null(space)) {
     obs <- unique(unlist(lapply(matches, names), use.names = FALSE))
-    if (is.null(space_name)) {
-      if (is.logical(space) && space) {
-        space <- select.lspace(terms = obs)$selected
-        space <- if (nrow(space)) rownames(space)[1] else NULL
-      }
-      if (is.character(space)) {
-        if (verbose) cat("loading space (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
-        space_name <- space
-        space <- lma_lspace(obs, space, dir = space_dir)
-      }
+    if (is.logical(space) && space) {
+      space <- select.lspace(terms = obs)$selected
+      space <- if (nrow(space)) rownames(space)[1] else NULL
+    }
+    if (is.character(space)) {
+      if (verbose) cat("loading space (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
+      space_name <- space
+      space <- lma_lspace(obs, space, dir = space_dir)
     }
     if (!nrow(space) || !any(obs %in% rownames(space))) space <- NULL
     if (is.null(space)) {
@@ -145,10 +158,10 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
           if (verbose) cat("parsing phrases (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
           split_parts <- strsplit(obs[phrase], "[ _/-]")
           parts <- unique(unlist(split_parts, use.names = FALSE))
-          part_vectors <- if (!is.null(space_name)) {
-            lma_lspace(parts, space_name)
-          } else {
+          part_vectors <- if (is.null(space_name)) {
             if (any(parts %in% rownames(space))) space[parts[parts %in% rownames(space)]] else space[0, ]
+          } else {
+            lma_lspace(parts, space_name)
           }
           if (nrow(part_vectors)) {
             space_terms <- rownames(part_vectors)
@@ -173,6 +186,7 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
   if (verbose) cat("preparing results (", round(proc.time()[[3]] - st, 4), ")\n", sep = "")
   terms <- cbind(terms, do.call(rbind, lapply(matches, function(m) {
     hits <- if (length(m)) {
+      if (collapsed_terms) m[!is.na(m) & m != 0] <- 1L
       if (has_space) {
         msim <- m
         if (length(m) == 1) {
@@ -188,7 +202,7 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
             msim[su] <- as.numeric(sims[, which.max(colMeans(sims))])
           }
         }
-        o <- order(msim, decreasing = TRUE)
+        o <- order(msim, -nchar(names(m)), decreasing = TRUE)
         m <- m[o]
         if (as_string) {
           paste(paste0(names(m), " (", if (!as_terms) paste0(m, ", "), round(msim[o], 2), ")"), collapse = ", ")
@@ -196,7 +210,7 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
           list(as.data.frame(rbind(m, msim[o])))
         }
       } else {
-        m <- sort(m, TRUE)
+        m <- m[order(m, -nchar(names(m)), decreasing = TRUE)]
         if (as_string) {
           paste(paste0(names(m), if (!as_terms) paste0(" (", m, ")")), collapse = ", ")
         } else {
@@ -208,11 +222,19 @@ report_term_matches <- function(dict, text = NULL, space = NULL, glob = TRUE,
       if (as_string) "" else list(data.frame())
     }
     if (!as_string) rownames(hits[[1]]) <- NULL
-    res <- data.frame(
-      count = sum(m),
-      max_count = if (all(is.na(m))) 0L else max(m, na.rm = TRUE),
-      variants = length(m)
-    )
+    res <- if (collapsed_terms) {
+      data.frame(
+        count = length(m),
+        max_count = if (all(is.na(m))) 0L else 1L,
+        variants = length(m)
+      )
+    } else {
+      data.frame(
+        count = sum(m),
+        max_count = if (all(is.na(m))) 0L else max(m, na.rm = TRUE),
+        variants = length(m)
+      )
+    }
     if (has_space) {
       res$space <- space_name
       if (all(is.na(msim))) {